Finalized add button functionality
Added a method when the user presses the
operation button again, it will calculate
the new sum out and if the user presses
the button again, the next math
expression will use the result
from the first expression. Also
removed the double button error handling.
Added an error that occurs when you press the
equals sign without choosing an operation.