Stopped operator buttons from double clicking
I finally figured out how to stop the user
from pressing the same operator button twice.
I added a counter variable that will count
how many times the user presses the
operator button, and upon pressing it more
than once, it will command the
operator buttons to not do anything. The
operator buttons ONLY evaluate when the user
presses the numbers buttons for the
second operand. Pressing the number buttons
also reset the operator button counter so that it
doesn't interfere with the normal
evaluation process.