Added error checking and functionality to sign
Added functionality to sign change button, where
the sign button splits the number into an array
and checks for a negative sign. If it exists, then
it will remove the negative sign with shift().
If it doesn't exist, then it will add one with
unshift("-"). Also added error checks for
operation buttons to prevent user from activating
operations while the calculator is error'd.