Added digit and operator keyboard support
Using an event listener that is attached to the
entire DOM (the document) that listens to
each "keydown" event. After that, we use a switch
case to handle all of the valid events
that we want, the digits and the operators.
When operators are clicked, simulate the "click"
event on them with click()