Finished backspace button javascript logic
Splits the current number into a number array,
then uses .pop() to remove the last element
the user entered in, then .joins() the number
back together to display to the user. Also
included backspace button keyboard support.
I believe this is the last feature I needed
to work on, so this should be the calculator
finished.